move i/o functions to other file
